ai-assistant/requirements.txt
Flask
Flask-Cors
gunicorn
numpy
IPython
python-dotenv
google-ai-generativelanguage